requirements.txt
pillow==8.0.0
django==4.0.1
django-ckeditor==6.2.0